+++ met-web/src/components/common/dragndrop/DragItem.tsx
@@ -1,0 +1,93 @@
+import React, { useRef } from 'react';
+import { Box } from '@mui/material';
+import type { Identifier, XYCoord } from 'dnd-core';
+import { useDrag, useDrop } from 'react-dnd';
+
+interface DragItem {
+    index: number;
+    testId?: string;
+    moveItem: (dragIndex: number, hoverIndex: number) => void;
+    children: React.ReactNode;
+}
+
+interface Draggable {
+    index: number;
+    id: string;
+    type: string;
+}
+
+const DragItem = ({ index, moveItem, testId, children }: DragItem) => {
+    const ref = useRef<HTMLDivElement>(null);
+
+    const [{ handlerId }, drop] = useDrop<Draggable, void, { handlerId: Identifier | null }>({
+        accept: 'Contact',
+        collect(monitor) {
+            return {
+                handlerId: monitor.getHandlerId(),
+            };
+        },
+        hover(item: Draggable, monitor) {
+            if (!ref.current) {
+                return;
+            }
+            const dragIndex = item.index;
+            const hoverIndex = index;
+
+            // Don't replace items with themselves
+            if (dragIndex === hoverIndex) {
+                return;
+            }
+
+            // Determine rectangle on screen
+            const hoverBoundingRect = ref.current?.getBoundingClientRect();
+
+            // Get vertical middle
+            const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
+
+            // Determine mouse position
+            const clientOffset = monitor.getClientOffset();
+
+            // Get pixels to the top
+            const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
+
+            // Only perform the move when the mouse has crossed half of the items height
+            // When dragging downwards, only move when the cursor is below 50%
+            // When dragging upwards, only move when the cursor is above 50%
+
+            // Dragging downwards
+            if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
+                return;
+            }
+
+            // Dragging upwards
+            if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
+                return;
+            }
+
+            // Time to actually perform the action
+            moveItem(dragIndex, hoverIndex);
+
+            // Note: we're mutating the monitor item here!
+            // Generally it's better to avoid mutations,
+            // but it's good here for the sake of performance
+            // to avoid expensive index searches.
+            item.index = hoverIndex;
+        },
+    });
+
+    const [{ isDragging }, drag] = useDrag({
+        type: 'Contact',
+        item: () => {
+            return { testId, index };
+        },
+        collect: (monitor: any) => ({
+            isDragging: monitor.isDragging(),
+        }),
+    });
+
+    drag(drop(ref));
+
+    return <Box ref={ref}>{children}</Box>;
+};
+
+export default DragItem;
