+++ met-web/src/components/engagement/form/EngagementWidgets/WhoIsListeningForm.tsx
@@ -6,12 +6,12 @@ import { useAppDispatch } from 'hooks';
 import { openNotification } from 'services/notificationService/notificationSlice';
 import { postWidgetItems } from 'services/widgetService';
 import { WidgetDrawerContext } from './WidgetDrawerContext';
-import ContantInfoPaper from './ContactInfoPaper';
 import { WidgetType } from 'models/widget';
 import update from 'immutability-helper';
+import ContactBlock from './ContactBlock';
 
 const WhoIsListeningForm = () => {
-    const { handleWidgetDrawerOpen, handleAddContactDrawerOpen, loadingContacts, contacts, widgets, clearSelected } =
+    const { handleWidgetDrawerOpen, handleAddContactDrawerOpen, loadingContacts, contacts, widgets } =
         useContext(WidgetDrawerContext);
     const dispatch = useAppDispatch();
     const [selectedContact, setSelectedContact] = useState<Contact | null>(null);
@@ -24,17 +24,6 @@ const WhoIsListeningForm = () => {
         );
         setAddedContacts(filteredContacts);
     }, [contacts]);
-
-    const moveContact = useCallback((dragIndex: number, hoverIndex: number) => {
-        setAddedContacts((prevContacts: Contact[]) =>
-            update(prevContacts, {
-                $splice: [
-                    [dragIndex, 1],
-                    [hoverIndex, 0, prevContacts[dragIndex]],
-                ],
-            }),
-        );
-    }, []);
 
     const widgetId = widgets.filter((widget) => widget.widget_type_id === WidgetType.WhoIsListening)[0]?.id || null;
 
@@ -49,12 +38,6 @@ const WhoIsListeningForm = () => {
         }
 
         setAddedContacts([...addedContacts, selectedContact]);
-    };
-
-    const removeContact = (contact: Contact) => {
-        const filteredContacts = addedContacts.filter((addedContact) => contact.id != addedContact.id);
-        clearSelected();
-        setAddedContacts(filteredContacts);
     };
 
     const addWidgetItems = async () => {
@@ -131,19 +114,9 @@ const WhoIsListeningForm = () => {
                     </SecondaryButton>
                 </Grid>
             </Grid>
-
-            {addedContacts.map((addedContact, index) => {
-                return (
-                    <Grid key={`added-contact-${addedContact.id}`} item xs={12}>
-                        <ContantInfoPaper
-                            moveContact={moveContact}
-                            index={index}
-                            removeContact={removeContact}
-                            contact={addedContact}
-                        />
-                    </Grid>
-                );
-            })}
+            <Grid item xs={12}>
+                <ContactBlock addedContacts={addedContacts} setAddedContacts={setAddedContacts} />
+            </Grid>
             <Grid item xs={12} container direction="row" spacing={1} justifyContent={'flex-start'} marginTop="8em">
                 <Grid item>
                     <PrimaryButton
